Add alternative that only integrates over the volume of scatterers
This treats scatteters as cubes instead of spheres, (which has to be changed, but it was easier to implement). Integrating over every scatterer explicitly should be possible because the integral is 0 outside of scatterers, so only their volume matters. For computational speed, nquad should be used, where a spherical integration should also be possible.